+++ src/textual/widgets/_radio_set.py
@@ -78,6 +78,9 @@ class RadioSet(Container, can_focus=True, can_focus_children=False):
         This message can be handled using an `on_radio_set_changed` method.
         """
 
+        ON_MATCHABLE_ATTRIBUTES = {"pressed"}
+        """Message attributes that can be used with the [`on` decorator][textual.on]."""
+
         def __init__(self, radio_set: RadioSet, pressed: RadioButton) -> None:
             """Initialise the message.