+++ src/textual/widgets/_tabbed_content.py
@@ -88,6 +88,9 @@ class TabbedContent(Widget):
     class TabActivated(Message):
         """Posted when the active tab changes."""
 
+        ON_MATCHABLE_ATTRIBUTES = {"tab"}
+        """Message attributes that can be used with the [`on` decorator][textual.on]."""
+
         def __init__(self, tabbed_content: TabbedContent, tab: Tab) -> None:
             """Initialize message.